+++ pkg/util/apptainerconf/config.go
@@ -128,6 +128,10 @@ type File struct {
 	DownloadPartSize    uint   `default:"5242880" directive:"download part size"`
 	DownloadBufferSize  uint   `default:"32768" directive:"download buffer size"`
 	SystemdCgroups      bool   `default:"yes" authorized:"yes,no" directive:"systemd cgroups"`
+	// apptheus unix socket
+	ApptheusSocketPath string `default:"/run/apptheus/gateway.sock" directive:"apptheus communication socket path"`
+	// Allow monitoring by apptheus, default is `no` because it requires an additional tool, i.e. apptheus
+	AllowMonitoring bool `default:"no" authorized:"yes,no" directive:"allow monitoring"`
 }
 
 // NOTE: if you think that we may want to change the default for any
@@ -530,4 +534,15 @@ download buffer size = {{ .DownloadBufferSize }}
 # Whether to use systemd to manage container cgroups. Required for rootless cgroups
 # functionality. 'no' will manage cgroups directly via cgroupfs.
 systemd cgroups = {{ if eq .SystemdCgroups true }}yes{{ else }}no{{ end }}
+
+# APPTHEUS SOCKET PATH: [STRING]
+# DEFAULT: /run/apptheus/gateway.sock
+# Defines apptheus socket path
+{{ if ne .ApptheusSocketPath "" }}apptheus socket path = {{ .ApptheusSocketPath }}{{ end }}
+
+# ALLOW MONITORING: [BOOL]
+# DEFAULT: no
+# Allow to monitor the system resource usage of apptainer. To enable this option
+# additional tool, i.e. apptheus, is required.
+allow monitoring = {{ if eq .AllowMonitoring true }}yes{{ else }}no{{ end }}
 `